+++ app/api_views.py
@@ -1,24 +1,12 @@
-# Local Python Modules
-import predictor
-import utils
-from elo_params import params
+# Module Imports
+from app import api,predictor,utils
+from app.elo_params import params
 
-# Other Modules
-from flask import Flask,render_template,request,jsonify,url_for,redirect,Blueprint
+# Package Imports
+from flask_restplus import Resource,reqparse
+from flask import render_template,request,jsonify,url_for,redirect
 import requests
 import json
-from flask_restplus import Api,Resource,reqparse
-
-app = Flask(__name__,
-                template_folder='templates',
-                static_folder='static')
-                
-app.config.from_pyfile('./config/gunicorn.conf.py')
-
-blueprint = Blueprint('api',__name__,url_prefix='/elo_api')
-api_name = 'Customer Loyalty Prediction Model'
-api = Api(blueprint,default=api_name,doc='/documentation')
-app.register_blueprint(blueprint)
 
 # API ROUTES
 @api.route("/sample")
@@ -93,33 +81,3 @@ class percentiles(Resource):
         pred = float(args['pred'])
         percentile = predictor.get_percentile(pred)
         return percentile 
-    
-# APP ROUTES
-@app.route("/")
-@app.route("/home")
-def home():
-    return render_template("home.html")
-
-@app.route("/demo")
-def demo():
-    sample_cols,sample_vals,metric_cols,metric_vals,preds,percentile = predictor.get_demo_values()
-    return render_template("demo.html",
-                            sample_cols=sample_cols,
-                            sample_vals=sample_vals,
-                            metric_cols=metric_cols,
-                            metric_vals=metric_vals,
-                            pred=preds,
-                            percentile=percentile)
-
-@app.route("/contact")
-def contact():
-    return render_template("contact.html")
-
-@app.route("/nav")
-def nav():
-    return render_template("nav_bar.html")
-
-# export FLASK_APP=app.py
-
-if __name__ == "__main__":
-    app.run()
