+++ app/elo_params.py
@@ -1,0 +1,21 @@
+import numpy as np 
+class params:
+    '''
+    Class Containing all variables needed to execute project
+    '''
+    def __init__(self):
+        '''
+        variables:
+
+        folder_path - relative path to relavent files
+        
+        '''
+        # Path to Folder
+
+        # Local Path
+        self.folder_path = "./app/static/data/"
+
+        # Heroku Path
+        # self.folder_path = "/app/static/data/"
+        
+        return
