+++ app/static/images/outliers_effect.png
