+++ app/templates/nav_bar.html
