--- config/gunicorn.conf.py
@@ -1,5 +1,0 @@
-def when_ready(server):
-    open('/tmp/app-initialized', 'w').close()
-
-
-bind = 'unix:///tmp/nginx.socket'
