+++ config/nginx.conf.erb
@@ -1,6 +1,5 @@
 daemon off;
-# Heroku dynos have at least 4 cores.
-worker_processes <%= ENV['NGINX_WORKERS'] || 4 %>;
+worker_processes auto;
 
 events {
   use epoll;
@@ -20,26 +19,21 @@ http {
   error_log logs/nginx/error.log;
 
   include mime.types;
+  default_type application/octet-stream;
   sendfile on;
 
   # Must read the body in 5 seconds.
-  client_body_timeout <%= ENV['NGINX_CLIENT_BODY_TIMEOUT'] || 5 %>;
-
-  upstream app_server {
-    server unix:/tmp/nginx.socket fail_timeout=0;
-  }
+  client_body_timeout 5;
 
   server {
     listen <%= ENV["PORT"] %>;
     server_name _;
     keepalive_timeout 5;
-    client_max_body_size <%= ENV['NGINX_CLIENT_MAX_BODY_SIZE'] || 1 %>M;
 
     location / {
-      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-      proxy_set_header Host $http_host;
-      proxy_redirect off;
-      proxy_pass http://app_server;
+      include uwsgi_params;
+      more_set_headers 'X-Request-Id: $http_x_request_id';
+      uwsgi_pass unix:/tmp/uwsgi.socket;
     }
   }
 }
