+++ run.py
@@ -1,0 +1,7 @@
+from app import app
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0",port=80)
+
+# export FLASK_APP=run.py
+# export FLASK_ENV=development
