+++ uwsgi.ini
@@ -1,0 +1,9 @@
+[uwsgi]
+wsgi-file = run.py
+callable = app
+socket = /tmp/uwsgi.socket
+master = true
+memory-report = true
+processes = 4
+threads = 20
+hook-accepting1 = exec:touch /tmp/app-initialized
